Move shuffle to player controller
